webapp/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
boto3==1.26.81
botocore==1.29.81
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
exceptiongroup==1.1.0
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.2
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
statsd==4.0.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0